move commands, flags, opsarguments to seperate file
